adapted authorized request to remove potential security risk

+++ CarHut/src/main/java/com/carhut/services/SavedSearchesService.java
@@ -5,8 +5,11 @@ import com.carhut.database.repository.UserCredentialsRepository;
 import com.carhut.enums.RequestStatusEntity;
 import com.carhut.models.carhut.SavedSearch;
 import com.carhut.models.security.User;
+import com.carhut.util.exceptions.authentication.CarHutAuthenticationException;
 import com.carhut.util.exceptions.savedsearches.SavedSearchesException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,7 +22,15 @@ public class SavedSearchesService {
     @Autowired
     private UserCredentialsRepository userCredentialsRepository;
 
-    public RequestStatusEntity addNewSavedSearch(SavedSearch savedSearch) throws SavedSearchesException {
+    public RequestStatusEntity addNewSavedSearch(SavedSearch savedSearch) throws SavedSearchesException, CarHutAuthenticationException {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userSecurityContextHolder = ((User)authentication.getPrincipal());
+
+        if (!userSecurityContextHolder.getUsername().equals(userCredentialsRepository.getUsernameByUserId(savedSearch.getUserId()))) {
+            throw new CarHutAuthenticationException("Unauthorized access add new saved search.");
+        }
+
         try {
             SavedSearch newSavedSearch = new SavedSearch(savedSearch.getUserId(), savedSearch.getSortBy(), savedSearch.getOffersPerPage(),
                     savedSearch.getPriceFrom(), savedSearch.getPriceTo(), savedSearch.getMileageFrom(),
@@ -34,7 +45,14 @@ public class SavedSearchesService {
     }
 
 
-    public List<SavedSearch> getSavedSearchesByUsername(String username) throws SavedSearchesException {
+    public List<SavedSearch> getSavedSearchesByUsername(String username) throws SavedSearchesException, CarHutAuthenticationException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User userSecurityContextHolder = ((User)authentication.getPrincipal());
+
+        if (!userSecurityContextHolder.getUsername().equals(username)) {
+            throw new CarHutAuthenticationException("Unauthorized access to get saved searches.");
+        }
+
         try {
             User user = userCredentialsRepository.findUserByUsername(username);
             return savedSearchesRepository.getSavedSearchesByUserId(user.getId());
@@ -46,8 +64,16 @@ public class SavedSearchesService {
     }
 
     public void removeSavedSearch(String id) throws SavedSearchesException {
+
         try {
             SavedSearch savedSearch = savedSearchesRepository.getSavedSearchesBySavedSearchId(id);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            User userSecurityContextHolder = ((User)authentication.getPrincipal());
+
+            if (!userSecurityContextHolder.getUsername().equals(userCredentialsRepository.getUsernameByUserId(savedSearch.getUserId()))) {
+                throw new CarHutAuthenticationException("Unauthorized access to remove saved searches.");
+            }
+
             savedSearchesRepository.delete(savedSearch);
         }
         catch (Exception e) {
